♻️ refactor: useGroupStore로 그룹 전역 상태 분리
